requirements.txt
loguru==0.7.2
gradio==3.50.2
transformers==4.36.0
soundfile==0.12.1
ffmpeg-python==0.2.0
# mac os
#--index-url https://download.pytorch.org/whl/cpu
#--extra-index-url https://pypi.org/simple
#torch==2.6.0

# Windows
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.1.2+cpu
accelerate==0.27.2
openai==1.3.0
python-dotenv==1.0.0
httpx==0.24.1